z2/requirements.txt
flask
mysql-connector-python
pdfkit
wkhtmltopdf